Added ORACLE environment variables required to start SQL Plus
Adding 'logoutput' attribute showed that SQL Plus was not being
initialized due to missing ORACLE environment variables. These same
environment variables were getting set in an interactive shell for the
user 'vagrant'.

Setting up the environment vairables manually is brittle, but works atm
and allows the setup.sql to run.

In a subsequent change, this will be replaced by running a script which:
1. Either does a sudo to vagrant from root, hopefull setting up the env
or,
2. Calls the oracle installation's setenv.sh to setup the env.